Add initial implementation of TinyG simulator.
TinyG firmware is compiled using the host gcc, and the most important I/O functions are either stubs or fakes.
For example, instead of reading from a serial port, the simulator reads from stdin. The output goes to stdout.

To build simulator: cd firmware/tinyg/sim && make
To run simulator: ./tinyg.elf
To debug simulator: gdb ./tinyg.elf

+++ firmware/tinyg/json_parser.c
@@ -485,7 +485,7 @@ uint16_t json_serialize(cmdObj_t *cmd, char_t *out_buf, uint16_t size)
 			}
 			else if (cmd->objtype == TYPE_DATA)	{
 				uint32_t *v = (uint32_t*)&cmd->value;
-				str += (char_t)sprintf((char *)str, "\"0x%lx\"", *v);
+				str += (char_t)sprintf((char *)str, "\"0x%lx\"", (unsigned long)*v);
 			}
 			else if (cmd->objtype == TYPE_STRING)	{ str += (char_t)sprintf((char *)str, "\"%s\"",(char *)*cmd->stringp);}
 			else if (cmd->objtype == TYPE_ARRAY)	{ str += (char_t)sprintf((char *)str, "[%s]",  (char *)*cmd->stringp);}
